A type mismatch in sound channel data messed up the sound after loading a saved state in certain games. int (32 bit) is now short (16 bit).

+++ source/soundux.h
@@ -213,8 +213,8 @@ typedef struct
     uint32 dummy [3];
     Channel channels [NUM_CHANNELS];
     // bool8 no_filter;
-    int master_volume [2]; /* range is -128 .. 127 */
-    int echo_volume [2]; /* range is -128 .. 127 */
+    short master_volume [2]; /* range is -128 .. 127 */
+    short echo_volume [2]; /* range is -128 .. 127 */
     int noise_hertz;
 } SSoundData;
